+++ src/Ventanas/ModificarHorario.java
@@ -5,10 +5,10 @@
  */
 package Ventanas;
 
-import Clases.Conexion;
-import Clases.Dia_Medico;
-import Clases.Hora_Medico;
-import Clases.Medico;
+import dao.Conexion;
+import dao.DiaMedicoDao;
+import dao.HoraMedicoDao;
+import dao.MedicoDao;
 import java.sql.CallableStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -882,7 +882,7 @@ public class ModificarHorario extends javax.swing.JDialog {
 
             int ID_MaxD = 0;
 
-            Dia_Medico.Agregar_Dia_Medico(ID_MaxM, "L");
+            DiaMedicoDao.Agregar_Dia_Medico(ID_MaxM, "L");
 
             try {
                 resultado = Conexion.consulta("Select Max(ID_Dia_Medico) from Dia_Medico");
@@ -897,7 +897,7 @@ public class ModificarHorario extends javax.swing.JDialog {
             String HoraInicio = (String) cmbDesde.getSelectedItem();
             String HoraFinal = (String) cmbHasta.getSelectedItem();
 
-            Hora_Medico.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
+            HoraMedicoDao.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
 
         }
 
@@ -908,7 +908,7 @@ public class ModificarHorario extends javax.swing.JDialog {
 
             int ID_MaxD = 0;
 
-            Dia_Medico.Agregar_Dia_Medico(ID_MaxM, "M");
+            DiaMedicoDao.Agregar_Dia_Medico(ID_MaxM, "M");
 
             try {
                 resultado = Conexion.consulta("Select Max(ID_Dia_Medico) from Dia_Medico");
@@ -923,7 +923,7 @@ public class ModificarHorario extends javax.swing.JDialog {
             String HoraInicio = (String) cmbDesde2.getSelectedItem();
             String HoraFinal = (String) cmbHasta2.getSelectedItem();
 
-            Hora_Medico.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
+            HoraMedicoDao.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
 
         }
 
@@ -934,7 +934,7 @@ public class ModificarHorario extends javax.swing.JDialog {
 
             int ID_MaxD = 0;
 
-            Dia_Medico.Agregar_Dia_Medico(ID_MaxM, "X");
+            DiaMedicoDao.Agregar_Dia_Medico(ID_MaxM, "X");
 
             try {
                 resultado = Conexion.consulta("Select Max(ID_Dia_Medico) from Dia_Medico");
@@ -949,7 +949,7 @@ public class ModificarHorario extends javax.swing.JDialog {
             String HoraInicio = (String) cmbDesde3.getSelectedItem();
             String HoraFinal = (String) cmbHasta3.getSelectedItem();
 
-            Hora_Medico.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
+            HoraMedicoDao.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
 
         }
 
@@ -960,7 +960,7 @@ public class ModificarHorario extends javax.swing.JDialog {
 
             int ID_MaxD = 0;
 
-            Dia_Medico.Agregar_Dia_Medico(ID_MaxM, "J");
+            DiaMedicoDao.Agregar_Dia_Medico(ID_MaxM, "J");
 
             try {
                 resultado = Conexion.consulta("Select Max(ID_Dia_Medico) from Dia_Medico");
@@ -975,7 +975,7 @@ public class ModificarHorario extends javax.swing.JDialog {
             String HoraInicio = (String) cmbDesde4.getSelectedItem();
             String HoraFinal = (String) cmbHasta4.getSelectedItem();
 
-            Hora_Medico.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
+            HoraMedicoDao.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
 
         }
 
@@ -986,7 +986,7 @@ public class ModificarHorario extends javax.swing.JDialog {
 
             int ID_MaxD = 0;
 
-            Dia_Medico.Agregar_Dia_Medico(ID_MaxM, "V");
+            DiaMedicoDao.Agregar_Dia_Medico(ID_MaxM, "V");
 
             try {
                 resultado = Conexion.consulta("Select Max(ID_Dia_Medico) from Dia_Medico");
@@ -1001,7 +1001,7 @@ public class ModificarHorario extends javax.swing.JDialog {
             String HoraInicio = (String) cmbDesde5.getSelectedItem();
             String HoraFinal = (String) cmbHasta5.getSelectedItem();
 
-            Hora_Medico.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
+            HoraMedicoDao.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
 
         }
 
@@ -1012,7 +1012,7 @@ public class ModificarHorario extends javax.swing.JDialog {
 
             int ID_MaxD = 0;
 
-            Dia_Medico.Agregar_Dia_Medico(ID_MaxM, "S");
+            DiaMedicoDao.Agregar_Dia_Medico(ID_MaxM, "S");
 
             try {
                 resultado = Conexion.consulta("Select Max(ID_Dia_Medico) from Dia_Medico");
@@ -1027,7 +1027,7 @@ public class ModificarHorario extends javax.swing.JDialog {
             String HoraInicio = (String) cmbDesde6.getSelectedItem();
             String HoraFinal = (String) cmbHasta6.getSelectedItem();
 
-            Hora_Medico.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
+            HoraMedicoDao.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
 
         }
 
@@ -1038,7 +1038,7 @@ public class ModificarHorario extends javax.swing.JDialog {
 
             int ID_MaxD = 0;
 
-            Dia_Medico.Agregar_Dia_Medico(ID_MaxM, "D");
+            DiaMedicoDao.Agregar_Dia_Medico(ID_MaxM, "D");
 
             try {
                 resultado = Conexion.consulta("Select Max(ID_Dia_Medico) from Dia_Medico");
@@ -1053,7 +1053,7 @@ public class ModificarHorario extends javax.swing.JDialog {
             String HoraInicio = (String) cmbDesde9.getSelectedItem();
             String HoraFinal = (String) cmbHasta9.getSelectedItem();
 
-            Hora_Medico.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
+            HoraMedicoDao.Agregar_Hora_Medico(ID_MaxD, HoraInicio, HoraFinal);
 
         }
