+++ src/dao/Conexion.java
@@ -1,4 +1,4 @@
-package Clases;
+package dao;
 
 import java.sql.*;
 import javax.swing.JOptionPane;
