+++ src/dao/EspecialidadDao.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Clases;
+package dao;
 
 import java.sql.CallableStatement;
 import java.sql.ResultSet;
@@ -14,7 +14,7 @@ import javax.swing.JOptionPane;
  *
  * @author UNI
  */
-public class Especialidad {
+public class EspecialidadDao {
 
     public static ResultSet resultado;
